Changed everything to preprocessor and it passed my tests.
