+++ src/mirror/models/mirror_llama_model.py
@@ -7,7 +7,7 @@ from typing import Literal
 from mirror.models.mirror_model import MirrorModel
 from mirror.models.model_util import build_causal_lm, IGNORE_ID
 from mirror.models.configuration_llama import LlamaConfig
-from mirror.tokenizers.mirror_llama_tokenizer import MirrorLlamaTokenizer
+from mirror.preprocessors.mirror_llama_preprocessor import MirrorLlamaPreprocessor
 from mirror.types import AttentionMaskBatch, Loss, TokenBatch, TokenTensor
 from mirror.util import pad_to_longest
 from mirror.row_types import TextRow
@@ -18,25 +18,19 @@ class MirrorLlamaModel(MirrorModel[TextRow, TokenTensor, tuple[TokenBatch, Atten
         initialization: Literal["3.2-1B", "3.2-1B-Instruct"] | LlamaConfig = "3.2-1B-Instruct"
     ) -> None:
         super().__init__()
-        default_tokenizer_hf_name = "meta-llama/Llama-3.2-1B-Instruct"
+        default_preprocessor_hf_name = "meta-llama/Llama-3.2-1B-Instruct"
 
         if isinstance(initialization, LlamaConfig):
             self.model = AutoModelForCausalLM.from_config(initialization)
-            self._tokenizer = MirrorLlamaTokenizer(default_tokenizer_hf_name)
+            self._preprocessor = MirrorLlamaPreprocessor(default_preprocessor_hf_name)
         else:
             hf_model_name = f"meta-llama/Llama-{initialization}"
             self.model = build_causal_lm(hf_model_name, weights="pretrained")
-            self._tokenizer = MirrorLlamaTokenizer(hf_model_name)
+            self._preprocessor = MirrorLlamaPreprocessor(hf_model_name)
 
     @property
-    def tokenizer(self) -> MirrorLlamaTokenizer:
-        return self._tokenizer
-
-    def preprocess_example(self, example: TextRow) -> TokenTensor:
-        return self.tokenizer.encode(example['text'])
-
-    def collate(self, examples: list[TokenTensor]) -> tuple[TokenBatch, AttentionMaskBatch]:
-        return pad_to_longest(examples, pad_token=self.tokenizer.pad_token_id)
+    def preprocessor(self) -> MirrorLlamaPreprocessor:
+        return self._preprocessor
 
     def training_step(self, batch: tuple[TokenBatch, AttentionMaskBatch]) -> Loss:
         input_ids, attention_mask = batch
