+++ src/mirror/models/mirror_model.py
@@ -2,25 +2,17 @@ from torch.optim import Optimizer
 import torch.nn as nn
 from abc import ABC, abstractmethod
 from mirror.types import Loss
-from mirror.tokenizers.mirror_tokenizer import MirrorTokenizer
+from mirror.preprocessors.mirror_preprocessor import MirrorPreprocessor
 
 
 class MirrorModel[RawT, ProcessedT, BatchT](ABC, nn.Module):
     @property
     @abstractmethod
-    def tokenizer(self) -> MirrorTokenizer:
+    def preprocessor(self) -> MirrorPreprocessor:
         pass
 
     @abstractmethod
     def training_step(self, batch: BatchT) -> Loss:
-        pass
-
-    @abstractmethod
-    def preprocess_example(self, example: RawT) -> ProcessedT:
-        pass
-
-    @abstractmethod
-    def collate(self, examples: list[ProcessedT]) -> BatchT:
         pass
 
     @abstractmethod
