+++ src/mirror/models/placeholder_model.py
@@ -3,9 +3,9 @@ import torch.optim as optim
 import torch.nn as nn
 
 from mirror.models.mirror_model import MirrorModel
-from mirror.tokenizers.placeholder_tokenizer import PlaceholderTokenizer
+from mirror.preprocessors.placeholder_preprocessor import PlaceholderPreprocessor
 from mirror.types import AttentionMaskBatch, Loss, TokenBatch, TokenTensor
-from mirror.util import get_device, pad_to_longest
+from mirror.util import get_device
 
 from mirror.row_types import TextRow
 
@@ -13,21 +13,15 @@ class PlaceholderModel(MirrorModel[TextRow, TokenTensor, tuple[TokenBatch, Atten
     def __init__(self) -> None:
         super().__init__()
         self.parameter = nn.Parameter(torch.tensor([0.0], device=get_device()))
-        self._tokenizer = PlaceholderTokenizer()
+        self._preprocessor = PlaceholderPreprocessor()
 
     @property
-    def tokenizer(self) -> PlaceholderTokenizer:
-        return self._tokenizer
+    def preprocessor(self) -> PlaceholderPreprocessor:
+        return self._preprocessor    
 
-    def preprocess_example(self, example: TextRow) -> TokenTensor:
-        return self._tokenizer.encode(example['text'])
-    
     def training_step(self, batch: tuple[TokenBatch, AttentionMaskBatch]) -> Loss:
         tokens, attention_mask = batch
         return self.parameter
-    
-    def collate(self, examples: list[TokenTensor]) -> tuple[TokenBatch, AttentionMaskBatch]:
-        return pad_to_longest(examples, pad_token=self.tokenizer.pad_token_id)
 
     def configure_optimizers(self) -> optim.Optimizer:
         return optim.AdamW(self.parameters())
