+++ src/mirror/preprocessors/mirror_llama_preprocessor.py
@@ -1,13 +1,13 @@
 import torch
 from transformers import PreTrainedTokenizerBase
 
-from mirror.tokenizers.mirror_tokenizer import MirrorTokenizer
-from mirror.tokenizers.tokenizer_util import load_hf_tokenizer
-from mirror.types import TokenTensor
-from mirror.util import get_device
+from mirror.preprocessors.mirror_preprocessor import MirrorPreprocessor
+from mirror.preprocessors.preprocessor_util import load_hf_tokenizer
+from mirror.types import TokenTensor, TokenBatch, AttentionMaskBatch
+from mirror.util import get_device, pad_to_longest
+from mirror.row_types import TextRow
 
-
-class MirrorLlamaTokenizer(MirrorTokenizer):
+class MirrorLlamaPreprocessor(MirrorPreprocessor):
     def __init__(self, hf_model_name: str) -> None:
         self._hf_model_name = hf_model_name
         self._tokenizer: PreTrainedTokenizerBase = load_hf_tokenizer(hf_model_name)
@@ -16,7 +16,7 @@ class MirrorLlamaTokenizer(MirrorTokenizer):
 
     @property
     def tokenization_id(self) -> str:
-        return "MirrorLlamaTokenizer"
+        return "MirrorLlamaPreprocessor"
 
     def encode(self, text: str) -> TokenTensor:
         ids = self._tokenizer.encode(text, add_special_tokens=True)
@@ -25,9 +25,11 @@ class MirrorLlamaTokenizer(MirrorTokenizer):
             ids = [eos, eos] if len(ids) == 0 else [*ids, eos]
         return torch.tensor(ids, device=get_device(), dtype=torch.long)
 
-    def decode(self, tokens: TokenTensor) -> str:
-        tokens = tokens.tolist()
-        return self._tokenizer.decode(tokens, skip_special_tokens=False)
+    def preprocess_example(self, example: TextRow) -> TokenTensor:
+        return self.encode(example['text'])
+
+    def collate(self, examples: list[TokenTensor]) -> tuple[TokenBatch, AttentionMaskBatch]:
+        return pad_to_longest(examples, pad_token=self.pad_token_id)
 
     @property
     def pad_token_id(self) -> int:
